change almost all int-s to uint64_t-s

Former-commit-id: 148e2ddc00f1ba0651bcd04807cbfea3c4c68a83

+++ src/kmer-index/bitarray.c
@@ -11,17 +11,17 @@ bitarray_t* create_bitarray(uint64_t n)
   return array;
 }
 
-void add_to_bitarray(bitarray_t* array, int value)
+void add_to_bitarray(bitarray_t* array, uint64_t value)
 {
 	array->values[value / 8] = array->values[value / 8] | ( 1 << (value % 8));
 }
 
-void delete_from_bitarray(bitarray_t* array, int value)
+void delete_from_bitarray(bitarray_t* array, uint64_t value)
 {
 	array->values[value / 8] = array->values[value / 8] & ~(1 << (value % 8));
 }
 
-int is_member(bitarray_t* array, int value)
+int is_member(bitarray_t* array, uint64_t value)
 {
 	return (array->values[value / 8] & (1 << (value % 8)));
 }
